+++ iam/role/base/main.tf
@@ -1,11 +1,12 @@
 # L1
 
 module "assume_role_policy" {
-  source       = "../../../iam/policy/assume_role"
-  for_each     = local.l1_map.enable_assume_role ? { this = {} } : {}
-  account_map  = local.l1_map.assume_role_account_map
-  service_list = local.l1_map.assume_role_service_list
-  std_map      = var.std_map
+  source        = "../../../iam/policy/assume_role"
+  for_each      = local.l1_map.enable_assume_role ? { this = {} } : {}
+  account_map   = local.l1_map.assume_role_account_map
+  service_list  = local.l1_map.assume_role_service_list
+  user_arn_list = local.l1_map.assume_role_user_arn_list
+  std_map       = var.std_map
 }
 
 resource "aws_iam_policy" "this_created_policy" {